+++ package.json
@@ -46,6 +46,7 @@
     "@angular/platform-browser-dynamic": "^6.0.2",
     "@angular/platform-server": "^6.0.2",
     "@angular/router": "^6.0.2",
+    "@tjoskar/assert": "^1.0.0",
     "@types/core-js": "^0.9.46",
     "@types/mocha": "^5.2.0",
     "@types/node": "^10.1.2",
@@ -60,6 +61,7 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^3.0.0",
     "lodash.isequal": "^4.5.0",
+    "marble-test": "^3.0.0",
     "mocha": "^5.2.0",
     "null-loader": "^0.1.1",
     "protractor": "^5.3.2",
