+++ src/util/__test__/css.util.test.ts
@@ -1,5 +1,5 @@
-import { is, isNot } from './helpers/assert';
-import { removeCssClassName, addCssClassName, hasCssClassName } from '../src/utils';
+import { is } from '@tjoskar/assert';
+import { removeCssClassName, addCssClassName, hasCssClassName } from '../css.util';
 
 describe('Utils function', () => {
     const element = {} as HTMLImageElement;
